OP-3214 - extracted logic for asset hierarchy tree to lib
